Refactor files for Instructor Student View
